@@ -1,80 +1,70 @@
 {
+  "name": "jolly-roger-web",
+  "private": true,
+  "version": "0.0.1",
+  "type": "module",
+  "engines": {
+    "node": ">= 12.17.0"
+  },
   "dependencies": {
-    "@curi/interactions": "^2.0.2",
-    "@curi/router": "^2.1.1",
-    "@curi/svelte": "^2.0.0-beta.6",
-    "@ethersproject/abi": "^5.0.9",
-    "@ethersproject/bignumber": "^5.0.12",
-    "@ethersproject/bytes": "^5.0.8",
-    "@ethersproject/contracts": "^5.0.8",
-    "@ethersproject/providers": "^5.0.17",
-    "@ethersproject/solidity": "^5.0.7",
-    "@ethersproject/wallet": "^5.0.7",
-    "@hickory/browser": "^2.1.0",
-    "@urql/core": "^1.12.3",
-    "@urql/devtools": "^2.0.2",
-    "graphql": "^15.3.0",
-    "jolly-roger-common": "workspace:*",
-    "named-logs": "0.0.13",
-    "named-logs-console": "0.0.14",
-    "svelte": "^3.24.0",
-    "web3w": "^0.2.42",
-    "web3w-torus-loader": "^0.2.4",
-    "web3w-walletconnect-loader": "^0.2.0",
-    "wonka": "^4.0.14"
+    "@ethersproject/abi": "^5.1.2",
+    "@ethersproject/abstract-provider": "^5.1.0",
+    "@ethersproject/bignumber": "^5.1.1",
+    "@ethersproject/bytes": "^5.1.0",
+    "@ethersproject/contracts": "^5.1.1",
+    "@ethersproject/providers": "^5.1.2",
+    "@ethersproject/solidity": "^5.1.0",
+    "@urql/core": "2.1.2",
+    "@urql/devtools": "^2.0.3",
+    "graphql": "^15.5.0",
+    "named-logs": "^0.1.0",
+    "web3w": "^0.3.1",
+    "web3w-torus-loader": "^0.3.1",
+    "web3w-walletconnect-loader": "^0.3.1",
+    "wonka": "^4.0.15"
   },
   "devDependencies": {
-    "@curi/types": "^2.0.2",
-    "@snowpack/plugin-dotenv": "^2.0.5",
-    "@snowpack/plugin-postcss": "^1.1.0",
-    "@snowpack/plugin-svelte": "^3.5.0",
-    "@snowpack/plugin-typescript": "^1.2.0",
-    "@snowpack/web-test-runner-plugin": "^0.2.0",
-    "@testing-library/svelte": "^3.0.0",
-    "@types/chai": "^4.2.13",
-    "@types/mocha": "^8.2.0",
-    "@types/snowpack-env": "^2.3.2",
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "@typescript-eslint/parser": "^4.14.0",
-    "@web/test-runner": "^0.12.0",
-    "autoprefixer": "^10.2.3",
-    "chai": "^4.2.0",
-    "cross-env": "^7.0.3",
-    "dotenv-cli": "^4.0.0",
-    "eslint": "^7.18.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-svelte3": "^3.0.0",
+    "@sveltejs/adapter-node": "next",
+    "@sveltejs/adapter-static": "^1.0.0-next.9",
+    "@sveltejs/kit": "next",
+    "@tailwindcss/jit": "^0.1.18",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
+    "autoprefixer": "^10.2.5",
+    "bech32": "^2.0.0",
+    "bn.js": "^5.2.0",
+    "cssnano": "^5.0.3",
+    "eslint": "^7.26.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-svelte3": "^3.2.0",
+    "hash.js": "^1.1.7",
     "ipfs-deploy": "^8.0.1",
     "ipfs-gateway-emulator": "^4.2.1-ipfs.1",
-    "npm-run-all": "^4.1.5",
-    "postcss": "^8.2.4",
-    "postcss-cli": "^8.3.1",
-    "prepare-for-web": "^0.1.2",
-    "prettier": "^2.2.1",
-    "prettier-plugin-svelte": "^2.1.0",
-    "snowpack": "^3.0.13",
-    "snowpack-plugin-folder2routes": "^0.1.1",
-    "snowpack-plugin-hmr-inject": "^0.1.5",
-    "snowpack-plugin-ipfs": "^0.1.5",
-    "snowpack-plugin-window-modules": "^0.1.2",
-    "svelte-hmr": "^0.12.9",
-    "svelte-preprocess": "^4.0.8",
-    "tailwindcss": "^2.0.2",
-    "tailwindcss-question-mark": "^0.1.0",
-    "typescript": "^4.0.0"
+    "jolly-roger-common": "workspace:*",
+    "postcss": "^8.2.15",
+    "postcss-load-config": "^3.0.1",
+    "prepare-for-web": "^0.1.5",
+    "prettier": "~2.3.0",
+    "prettier-plugin-svelte": "^2.3.0",
+    "svelte": "^3.38.2",
+    "svelte-preprocess": "^4.7.3",
+    "sveltejs-adapter-ipfs": "^0.1.4",
+    "tailwindcss": "^2.1.2",
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4",
+    "vite": "^2.3.3"
   },
   "scripts": {
-    "prepare": "node ./.setup.js && prepare-for-web",
-    "lint": "eslint \"./**/*.{js,ts}\"",
-    "lint:fix": "eslint --fix \"./**/*.{js,ts}\"",
-    "format": "prettier --write \"./**/*.{ts,js,svelte}\"",
-    "dev:reload": "snowpack dev --reload",
-    "dev": "snowpack dev",
-    "build": "snowpack build",
-    "test": "web-test-runner \"src/**/*.test.ts\"",
+    "prepare": "node ./.setup.cjs && prepare-for-web --target static --targetFile src/app.html",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "format": "prettier --check .",
+    "format:fix": "prettier --write .",
+    "dev": "svelte-kit dev",
+    "build": "svelte-kit build",
+    "start": "svelte-kit start",
     "serve": "ipfs-emulator --only -d build -p 8080",
-    "build:serve": "run-s build serve",
-    "deploy": "dotenv -- ipfs-deploy -p pinata -s jolly-roger-snowpack build",
+    "deploy": "dotenv -- ipfs-deploy -p pinata -s jolly-roger build",
     "build:deploy": "run-s build deploy"
   }
 }
