@@ -1,9 +1,19 @@
-const autoPreprocess = require('svelte-preprocess');
+import preprocess from 'svelte-preprocess';
+import adapter_ipfs from 'sveltejs-adapter-ipfs';
 
-module.exports = {
-  preprocess: autoPreprocess({
-    defaults: {
-      script: 'typescript',
-    },
-  }),
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+  preprocess: preprocess(),
+
+  kit: {
+    adapter: adapter_ipfs({
+      removeBuiltInServiceWorkerRegistration: true,
+      injectPagesInServiceWorker: true,
+      injectDebugConsole: true,
+    }),
+    target: '#svelte',
+    trailingSlash: 'ignore',
+  },
 };
+
+export default config;
