@@ -1,14 +1,13 @@
 'use strict';
-require('dotenv').config();
-process.env.TS_NODE_FILES = 'true';
+process.env.TS_NODE_FILES = true;
 module.exports = {
   'allow-uncaught': true,
   diff: true,
   extension: ['ts'],
   recursive: true,
   reporter: 'spec',
-  require: ['ts-node/register', 'hardhat/register'],
-  slow: 75,
+  require: ['ts-node/register', 'hardhat/register'], // ['ts-node/register/transpile-only'], (for yarn link <plugin>)
+  slow: 300,
   spec: 'test/**/*.test.ts',
   timeout: 20000,
   ui: 'bdd',
