@@ -1,21 +1,15 @@
 import 'dotenv/config';
 import {HardhatUserConfig} from 'hardhat/types';
 import 'hardhat-deploy';
-import 'hardhat-deploy-ethers';
+import '@nomiclabs/hardhat-ethers';
 import 'hardhat-gas-reporter';
-import 'hardhat-typechain';
+import '@typechain/hardhat';
 import 'solidity-coverage';
 import {node_url, accounts} from './utils/network';
 
 const config: HardhatUserConfig = {
   solidity: {
-    version: '0.7.1',
-    settings: {
-      optimizer: {
-        enabled: true,
-        runs: 2000,
-      },
-    },
+    version: '0.7.6',
   },
   namedAccounts: {
     deployer: 0,
@@ -23,11 +17,29 @@ const config: HardhatUserConfig = {
   },
   networks: {
     hardhat: {
-      accounts: accounts(),
+      // process.env.HARDHAT_FORK will specify the network that the fork is made from.
+      // this line ensure the use of the corresponding accounts
+      accounts: accounts(process.env.HARDHAT_FORK),
+      forking: process.env.HARDHAT_FORK
+        ? {
+            url: node_url(process.env.HARDHAT_FORK),
+            blockNumber: process.env.HARDHAT_FORK_NUMBER
+              ? parseInt(process.env.HARDHAT_FORK_NUMBER)
+              : undefined,
+          }
+        : undefined,
     },
     localhost: {
       url: node_url('localhost'),
       accounts: accounts(),
+    },
+    staging: {
+      url: node_url('rinkeby'),
+      accounts: accounts('rinkeby'),
+    },
+    production: {
+      url: node_url('mainnet'),
+      accounts: accounts('mainnet'),
     },
     mainnet: {
       url: node_url('mainnet'),
@@ -44,10 +56,6 @@ const config: HardhatUserConfig = {
     goerli: {
       url: node_url('goerli'),
       accounts: accounts('goerli'),
-    },
-    staging: {
-      url: node_url('rinkeby'),
-      accounts: accounts('rinkeby'),
     },
   },
   paths: {
@@ -67,6 +75,16 @@ const config: HardhatUserConfig = {
   mocha: {
     timeout: 0,
   },
+  external: process.env.HARDHAT_FORK
+    ? {
+        deployments: {
+          // process.env.HARDHAT_FORK will specify the network that the fork is made from.
+          // these lines allow it to fetch the deployments from the network being forked from both for node and deploy task
+          hardhat: ['deployments/' + process.env.HARDHAT_FORK],
+          localhost: ['deployments/' + process.env.HARDHAT_FORK],
+        },
+      }
+    : undefined,
 };
 
 export default config;
