@@ -10,8 +10,12 @@ const messages = [
   'नमस्ते',
 ];
 
-function waitFor<T>(p: Promise<{wait: () => Promise<T>}>): Promise<T> {
-  return p.then((tx) => tx.wait());
+async function waitFor<T>(p: Promise<{wait: () => Promise<T>}>): Promise<T> {
+  const tx = await p;
+  try {
+    await ethers.provider.send('evm_mine', []); // speed up on local network
+  } catch (e) {}
+  return tx.wait();
 }
 
 async function main() {
