@@ -1,6 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-export function getParamsFromURL(str?: string): Record<string, string> {
-  const url = str || window.location.href;
+export function getParamsFromURL(url: string): Record<string, string> {
+  if (!url) {
+    return {};
+  }
   const obj: Record<string, string> = {};
   const hash = url.lastIndexOf('#');
 
@@ -22,7 +24,18 @@ export function getParamsFromURL(str?: string): Record<string, string> {
   return obj;
 }
 
-export function getParamsFromURLHash(str?: string): Record<string, string> {
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export function getParamsFromLocation(): Record<string, string> {
+  if (typeof window === 'undefined') {
+    return {};
+  }
+  return getParamsFromURL(window.location.href);
+}
+
+export function getHashParamsFromLocation(str?: string): Record<string, string> {
+  if (typeof window === 'undefined') {
+    return {};
+  }
   const url = str || window.location.hash;
   const obj: Record<string, string> = {};
   const hash = url.lastIndexOf('#');
@@ -39,7 +52,19 @@ export function getParamsFromURLHash(str?: string): Record<string, string> {
   return obj;
 }
 
+export function queryStringifyNoArray(query: Record<string, string>): string {
+  let str = '';
+  for (const key of Object.keys(query)) {
+    const value = query[key];
+    str += `${str === '' ? '?' : '&'}${key}=${value}`;
+  }
+  return str;
+}
+
 export function rebuildLocationHash(hashParams: Record<string, string>): void {
+  if (typeof window === 'undefined') {
+    return;
+  }
   let reconstructedHash = '';
   Object.entries(hashParams).forEach((param) => {
     if (reconstructedHash === '') {
@@ -76,9 +101,7 @@ async function chrome76Detection(): Promise<boolean> {
 }
 
 function isNewChrome(): boolean {
-  const pieces = navigator.userAgent.match(
-    /Chrom(?:e|ium)\/([0-9]+)\.([0-9]+)\.([0-9]+)\.([0-9]+)/
-  );
+  const pieces = navigator.userAgent.match(/Chrom(?:e|ium)\/([0-9]+)\.([0-9]+)\.([0-9]+)\.([0-9]+)/);
   if (pieces === null || pieces.length !== 5) {
     return false;
   }
@@ -128,10 +151,7 @@ export function isPrivateWindow(): Promise<boolean | null> {
         navigator.userAgent.includes('msie')
       ) {
         // Edge or IE
-        if (
-          !window.indexedDB &&
-          (window.PointerEvent || window.MSPointerEvent)
-        ) {
+        if (!window.indexedDB && (window.PointerEvent || window.MSPointerEvent)) {
           resolve(true);
         }
         resolve(false);
@@ -141,9 +161,7 @@ export function isPrivateWindow(): Promise<boolean | null> {
           resolve(chrome76Detection());
         }
 
-        const fs =
-          (window as any).RequestFileSystem ||
-          (window as any).webkitRequestFileSystem;
+        const fs = (window as any).RequestFileSystem || (window as any).webkitRequestFileSystem;
         if (!fs) {
           resolve(null);
         } else {
